requirements.txt
ollama==0.1.6
python-dotenv